+++ Assets/Plugins/StreamChat/Core/Requests/PaginationParamsRequest.cs
@@ -4,17 +4,17 @@ namespace StreamChat.Core.Requests
 {
     public partial class PaginationParamsRequest : RequestObjectBase, ISavableTo<PaginationParamsRequestDTO>
     {
-        public double? IdGt { get; set; }
+        public int? IdGt { get; set; }
 
-        public double? IdGte { get; set; }
+        public int? IdGte { get; set; }
 
-        public double? IdLt { get; set; }
+        public int? IdLt { get; set; }
 
-        public double? IdLte { get; set; }
+        public int? IdLte { get; set; }
 
-        public double? Limit { get; set; }
+        public int? Limit { get; set; }
 
-        public double? Offset { get; set; }
+        public int? Offset { get; set; }
 
         PaginationParamsRequestDTO ISavableTo<PaginationParamsRequestDTO>.SaveToDto() =>
             new PaginationParamsRequestDTO
